+++ Modules/Intent.Modules.Modelers.Domain.Events/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.4.0
+
+- Improvement: Added `[Processing Handler]` trait to `Domain Event Handler` elements.
+
 ### Version 4.3.0
 
 - Improvment: Added `Suggestion` (4.4.x feature) to auto-create a `Domain Event Handler` element from the `Domain Event`.